# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LC Management API",
        "description": "LC Management System API - Complete system for managing contracts, buyers, orders, and shipments. Includes contract management, order processing with cost tracking, shipment tracking with summary statistics, and comprehensive reporting capabilities.",
        "contact": {
            "name": "API Support",
            "email": "support@lcmanagement.example.com"
        },
        "version": "1.2.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server (Alternative)"
        }
    ],
    "paths": {
...
        "/api/banks": {
            "get": {
                "tags": [
                    "Banks"
                ],
                "summary": "List all active banks",
                "description": "Get list of banks with optional search and status filter",
                "operationId": "caa138ee4cdf9701a3ec3c3198e9f777",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by bank status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "active",
                            "enum": [
                                "active",
                                "inactive"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by name, SWIFT code, or country",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/banks/{id}": {
            "get": {
                "tags": [
                    "Banks"
                ],
                "summary": "Get bank details",
                "description": "Get detailed information about a specific bank",
                "operationId": "a449b3fe60623680a294ad750ef03c8e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bank not found"
                    }
                }
            }
        },
...
    },
...
}